Create different files for Diff, TableView and CollectionView tests
